refactor all existing js files to typescript
