+++ client/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@types/styled-components": "^5.1.7",
     "ipfs-http-client": "^49.0.2",
     "react": "16.11.0",
     "react-dom": "16.11.0",
