+++ client/src/components/Main/index.ts
@@ -1,0 +1,3 @@
+import Main from './Main';
+
+export default Main;
